fix assessment.set_data for unencrypted saving
when encryption was introduced in
[this commit](https://github.com/emoncms/MyHomeEnergyPlanner/commit/9a734cea63e8063663e163b2cae6462ba7f7e3c1)
it seems a regression was introduced for the *unencrypted* case

before (works):

```
$stmt->bind_param("ssi", $data, $mdate, $id);
```

after (fails):

```
$stmt->bind_param("sbi", $data, $mdate, $id);
```

from [the docs](https://www.php.net/manual/en/mysqli-stmt.bind-param.php) we see
that `s` means `string` and `b` means `blob`.

note: it's not clear to me why `mdate` (`array('type' => 'int(11)')`) should
be treated as a string, but that's how it was before, and I've tested
that `ssi` works when `sbi` does not.