Work around a bug in whitenoise
I'm adding a load of images that are probably unused in the app because
whitenoise complains about them being referenced but not existing, which
is fair enough.

I've also removed a line of a comment from the minified jquery-ui CSS
theme which contained text like this:
  url(%22images%2Fui-icons_cc0000_256x240.png%22)

Before I removed this, Whitenoise was trying to look for
  jquery-ui-1.12.1/"images/ui-icons_cc0000_256x240.png"
which obviously doesn't exist.

This will be fixed in future by a Django pull request:
  https://github.com/django/django/pull/11241