add AssessmentQuerySetMixin
This mixin defines the `get_queryset` function, ensuring that it
returns all assessments the user owns, in addition to all
assessments in the organisations the user is a member of.

This required an update to the tests, as these were previously
relying on the `permission_class` `IsMemberOfConnectedOrganisation`
to guard against viewing an assessment you weren't in the
organisation. Since this is now handled by `get_queryset` it
returns 404, rather than 403.