Rework walls table & display
Someone (I can't remember who now!) mentioned there were some usability
issues on the fabric elements page, so I took a look.  My first draft
of this work was not done in a modular fashion, so this commit is part
of a bigger piece of work that I've been splitting up and cleaning up.

My work here is somewhat exploratory, because I don't know the code well
enough yet, and so I'm not sure what new concepts need introducing into
it.  As such, some of the changes are adding new patterns into the code
that are not used elsewhere, and will probably need to be revisited
again in the future.

The key thing, though, is that I'm trying to make all aspects of the
code more readable and more explicit.  I think I'm succeeding.  Details:

BUGS DISCOVERED:
Up and down arrows are disabled because they were hiding a data loss bug

Visually:
- Make better use of the space available, avoid horizontal scrolling
- Put the totals in a table footer under the appropriate columns
- Use table heads/footers in a different colour
- Right-align numbers

In code:
- Use HTML5's <template> tag instead of our own custom thing
- Space out the fabric element HTML so that it's readable
- Use data-key=<key> as a template to set the real key= attribute
- Use data-section=<name> as a way to show and hide things easily
- Use data-action to tag buttons as needing certain attributes set for
  the library manager to work
- Not super happy with some of the JS right now but still better than
  before I think