Total rework of reporting code
Ongoing work in progress - this is the substantial base but various
things don't work properly after the changes yet.

Use nunjucks for templating, so that we can store the report in the
database instead of in Git - getting around the liability issues of
distributing the Carbon Co-op report.

We pass through a context object to nunjucks which gives us another
layer of indirection between the data model and the template.  It means
we can make potentially incompatible changes to the data model without
breaking reports, because we can shim something else in.

We are also displaying the report in a separate iframe now too - this
means that we can remove the 'print mode' hack.  It keeps the styling of
the report and the UI totally separate which means we can produce much
better quality reports with better CSS - rules don't need to be prefixed
with #report everywhere now, which also avoids issues with specificity.