Improve factoring of TargetBar
The Frontend Masters React course said not to have more than one
component in any given file but it seems ridiculous to have a separate
file for sub-components that are functionally inseparable from the main
one.  So I think this is worth documenting somewhere in our React style
guide (so far nonexistent).

So, split out a Target component instead of having drawLine, which was
the pre-React name.