Move the display of all 'target bars' to a single component
This approach is working out OK so far, though every time the graphs
are updated React is re-rendering the whole HTML subtree because we're
not using any React state mechanisms.

One wart is the 'width' parameter.  It's unpleasant to pass this in,
will try to fix this later.

Create a new directory, `data`, which will hold datasets in hopefully
smaller chunks than the existing `model-datasets.js` file.  Right now
it just has the target bar target data in it, in the format required
for display - so no more transformation needs to be done to it.