Create PageHeader component to replace 'topgraphic' subview
The React component is designed to have a clean external API, with the
data it needs being passed into it.  And this means collecting the data
from the outside and passing it through, and then passing it through
again whenever it changes.

To get the data out into React when a redraw is required, we pass in
a callback which the React component uses to pass us back the various
`setState` functions.  It's not pretty but it works, and I'm sure
there's a better way to do it, but I suspect that might be when more of
the whole app is using React.

It's nice at least to have a fairly clean API for manipulating the page
header from the non-React side.

Not done:
* There's some other tickets open around improving the page header, what
  bars are displayed etc.