Improve PageHeader
The abomination I introduced using a callback to pass the setState
functions back into the surrounding code is gone!  The reason I did this
was because I thought that calling ReactDOM.render repeatedly as a way
to redraw a component was not correct.  I got this impression from
some errors in the JavaScript console, but it turns out React is plenty
smart enough to handle this - it's the correct way to redraw components
from outside Reactland.

This commit also introduces a difference between a page that *never*
shows the graphics and the user choosing not to display the graphics.
In the former case, we don't even provide a 'show house graphic' button.
And in the latter case, we preserve the user's choice of shown/hidden
across different views.