Use accessor function to iterate over scenario IDs
Previously to this commit, it was impossible to include data in the
assessment that wasn't part of a particular scenario, because the keys
of the assessment data were assumed to all be scenario IDs.

With an accessor function, get_scenario_ids(), we can put non-scenario
data in the JSON blob if we prefix it with an underscore.  I'm doing
this now because there are definitely various bits of data that don't
belong in a scenario and are currently awkwardly placed on the baseline
scenario.  For example, the data from the commentary page got placed
very deep in project.master.household.commentary_* and also on each
individual scenario, but this is a PITA to work with so I want to move
it all to project._commentary.