Reactify the 'commentary' view
The main change here is shifting rendering the commentary page over to
React, but most of the churn is in creating infrastructure to make this
read and work nicely.  I don't know if the route I'm taking is the right
one but I remain hopeful.

Assessment
----------

So the Assessment class is a couple of classes plus a function that
creates us some proxies.

The concept here is that Assessment is a refactoring shim that may
eventually turn into the main calculation engine as it aggregates
behaviour to itself.  At the moment it's a data access API.

It takes the assessment data as received from the server (which in the
non-transpiled portion of the code is accessible as the global variable
`p`, not to be confused with the global variable `project`,
which === `p.data`) and a function which when called should re-run the
model and sync to the server (which is `update` in the non-transpiled
code).

Commentary
----------

A word on the constructor: at the moment, the startup phase of the
code runs the model to initialise all the fields by checking against
`undefined` before using the data and then filling it in.  But it's nice
if we could guarantee default values for data before we try to display
or use them.  This way we never need to check if things are undefined
before using them.

properties
----------

This is a handy way to avoid writing getters and setters for loads of
different fields; there will be a lot of them.

For some the simpler views having a class per view with a call to
properties() should be enough.

Commentary (view)
-----------------

The view itself is pretty straightforward I think, given the above
changes.  I don't know if it's the right move to include the label and
notes in the LongTextField (or NumberField) component or to move it
outside... I don't have any experience of how to factor these things
well yet.