Remove duplicate 'use solar hot water' toggles
The data was stored in two places:
- scenario.use_SHW
- scenario.water_heating.solar_water_heating

They were kept in sync using dubious methods anyway.  I've also removed
the toggle on the lefthand menu for the whether to use hot water calcs
as I think it likely didn't work anyway - it referenced data.use_SHW
which means it would have set the current scenario's use_SHW value.
rather then the scenario it's meant to be for in the menu.

The reason I'm doing this is because it gets in the way of getting the
tests working, though it has the handy side effects of removing a wart
from all over the code.