Remove double-layering of permissions
Previously, each individual assessment CBV was duplicating the
permission logic: the queryset was defined via AssessmentQuerySetMixin
as the set of assessments the user had permission to view, but each
CBV also wrote the same logic as DRF permissions classes.

This had resulted in some inconsistency: the change to only show the
user all assessments in an organisation if they are an admin of that
organisation (ff9ed087 / 5c3efbb5 / #495) was only implemented in the
view permissions, not in the queryset.

This commit fixes that.  Rely entirely on the queryset for access
control as this can be used in more places and doesn't require the
repetition.  And fix the queryset to reflect the access control only
previously provided at the DRF layer.