Give people more sensible usernames
At the moment it's whatever Auth0 sends us as a 'nickname', which tends
to just be the first segment of the email address.  We also weren't
saving the email address – so save that too.  Also, use the 'full name'
field, drop the first+last name charade.

After this change, users will have their email addresses and full names
saved in Django when before they had neither.