Switch from using setUpClass() to setUpTestData() and/or setUp()
Don't use setUpClass for creating Django models in a TestCase; this is
run outside database transactions and screws up future state.  I don't
know why this wasn't causing failures before but it is now, after the
requirements upgrade.  The pytest-django module looks like it got some
changes to test ordering so that might have been it.

Instead, use setUpTestData where the data will not be modified and setUp
where the data will be modified.  This is because any alterations to a
model created in setUpTestData() will persist between tests in the class
(though not beyond it).

In Django 3.2 this division can be undone and we can use setUpTestData
for everything because it will do some magic around modifications, what
a beaut.