Use people's full names instead of auto-generated usernames (#677)
This is just for the assessment view.  It doesn't apply to the
organisation editor, which can be fixed separately.

Provide id, name and email for a user in the assessment metadata from
the API.  Previously the ID and the username were provided.

Also switch to using a serializer for the user data rather than a mixin.