Fix the SelectField type errors
The problem before was that if you put the data you are trying use as a
value into the DOM and get it out again, you lose all the type info:
whatever you put in there, you get a 'string' back.  And while we as
smart humans know this, TS doesn't have a clue.  However, if you use the
index into the options array as the value in the <option>, you can use
that to get the right data back out again, and TS knows exactly what
type it is.