Fix report output re aesthetics booleans
This is a badly designed bit of data structure (and entirely my fault).
We should be exposing the aesthetics data as can_change_internal or
similar because it's completely intuitive what this *should* mean.
However, the output format of get_context_data() is the one reporty
thing that's being kept constant while everything else changes - so
let's not edit this now but come back to it when the next big report
change happens.