Significantly improve performance of library list API endpoint
Previously I was noticing 70+ database queries for a single request in
production, which was mostly the same query repeated. The number of
queries was O(n) where n was the number of libraries a user has access
to via an organisational membership.  The number of queries is now O(1),
or more concretely, 6.

- MyLibrariesMixin.my_libraries is refactored and runs only one query.
- ListCreateLibraries.get_queryset() prefetches various related fields
  so it only has to be done once.
- check_library_share_permissions() now makes use of this prefetched
  data instead of fetching it anew for each library