server/graphs: always produce keys for bar charts if given category labels
We are using a template which assumes that the 'fuel use' graph always
has a key. Up until this point, it always did - we just had the first
assessment through which only has electricity, even in the baseline
scenario. This meant that although category labels were specified, no
key was generated.

We should always generate a key if there are category labels for the
sake of determinism, and we can then leave the logic about whether to
display that key or not to the template.