client/ui/fabric/mutator: measures should always have cost_total and quantity included (#880)
I'm not sure what I was thinking here, to not include the specifics on
the measures output. I think I was trying to maintain a better schema
by avoiding duplicating data unnecessarily - but I think now is not the
time to fix the saved data representation. It's a mess and there are
dependencies on it.