client/ui/fabric: fix bug where lofts end up in walls section (#909)
All inserted and measure'd wall-likes were given the 'type' of
'external wall', so they ended up in the walls section, even if they
were e.g. lofts.

In fixing this, I noticed that a type guard was previously giving us
faulty type info. Fixing this led to a bit more logic in the library
selector to unify the different wall-likes into one shape of data,
and extends the 'EWI' flag to all wall-likes for simplicity.